Adding last do functions to do_util file from easyaccess
